+++ test/native/test_main.cpp
@@ -10,6 +10,8 @@ void runYaoTests(TestRunner& runner);
 void runHexagramTests(TestRunner& runner);
 void runImuTests(TestRunner& runner);
 void runMicrophoneResearchTests(TestRunner& runner);
+void runPowerControllerTests(TestRunner& runner);
+void runPowerManagerTests(TestRunner& runner);
 void runTransformTests(TestRunner& runner);
 void runUiLayoutTests(TestRunner& runner);
 void runOrderingTests(TestRunner& runner);
@@ -30,6 +32,8 @@ int main() {
   runHexagramTests(runner);
   runImuTests(runner);
   runMicrophoneResearchTests(runner);
+  runPowerControllerTests(runner);
+  runPowerManagerTests(runner);
   runTransformTests(runner);
   runUiLayoutTests(runner);
   runOrderingTests(runner);